refactor(agentgateway): rename package + Target interface; drop teleport
Rename the translator package to agentgateway and lean its public surface
on the domain it produces:

- package: internal/reconciler/translator → internal/reconciler/agentgateway
- Model → Config, Transform → Compile

Replace the Backend tagged-pointer struct with a sum type:

- Backend{Name, Target}
- Target interface with Kind() TargetKind
- HTTPTarget (carries Protocol: streamable-http | sse) collapses the
  former StreamableHTTP and SSE variants — both transports already share
  host/port/path.
- StdioTarget unchanged.

Rename AuthnConfig → Authn, TokenExchangeAuthn → TokenExchange.

Remove teleport authentication end-to-end in this package: AuthnTypeTeleport,
TeleportAuthn, the auth.Teleport branch in Compile, parseAuthnType acceptance,
and all teleport test cases. The MCPServerAuth.Teleport CRD field is
removed in a separate change.

100% coverage on the new package; golangci-lint clean.
