+++ internal/reconciler/agentgateway/doc.go
@@ -1,0 +1,4 @@
+// Package agentgateway compiles an MCPServer spec into a Config: an
+// agentgateway configuration independent of the eventual serialization
+// (Kubernetes CRDs or agw native YAML). Compile is pure.
+package agentgateway
